# docs/swagger.yaml
basePath: /
definitions:
  models.Billboard:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      label:
        type: string
      store_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Cart:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.CartItem:
    properties:
      base_price:
        type: number
      cart:
        $ref: '#/definitions/models.Cart'
      cart_id:
        type: string
      created_at:
        type: string
      discount_amount:
        type: number
      discount_percent:
        type: integer
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      tax_amount:
        type: number
      tax_percent:
        type: integer
      total_price:
        type: number
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      billboard:
        $ref: '#/definitions/models.Billboard'
      billboard_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Color:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
...
  models.Image:
    properties:
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  models.Order:
    properties:
      address:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: string
      is_paid:
        type: boolean
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      phone:
        type: integer
      store:
        $ref: '#/definitions/models.Store'
      store_id:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order:
        $ref: '#/definitions/models.Order'
      order_id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Product:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      color:
        $ref: '#/definitions/models.Color'
      color_id:
        type: string
      created_at:
        type: string
      discount_percent:
        type: integer
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      is_archived:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      price:
        type: number
      price_discount:
        type: number
      size:
        $ref: '#/definitions/models.Size'
      size_id:
        type: string
      stock:
        type: integer
      store_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Size:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  models.SizeCreate:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  models.SizeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  models.SizeUpdate:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  models.Store:
    properties:
      billboards:
        items:
          $ref: '#/definitions/models.Billboard'
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      colors:
        items:
          $ref: '#/definitions/models.Color'
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      sizes:
        items:
          $ref: '#/definitions/models.Size'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: integer
      stores:
        items:
          $ref: '#/definitions/models.Store'
        type: array
      updated_at:
        type: string
    type: object
...
  web.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: hi.boedi8@gmail.com
    name: hi.boedi8@gmail.com
    url: https://www.github.com/hiboedi
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/{storeId}/sizes:
    get:
      consumes:
      - application/json
      description: FindAll Size from the store
      parameters:
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SizeResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: FindAll Size from the store
      tags:
      - Size
    post:
      consumes:
      - application/json
      description: create Size for the store
      parameters:
      - description: Size create
        in: body
        name: Size
        required: true
        schema:
          $ref: '#/definitions/models.SizeCreate'
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SizeResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: create Size for the store
      tags:
      - Size
  /api/{storeId}/sizes/{sizeId}:
    delete:
      consumes:
      - application/json
      description: Delete Size from the store
      parameters:
      - description: Size ID
        in: path
        name: sizeId
        required: true
        type: string
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Delete Size from the store
      tags:
      - Size
    get:
      consumes:
      - application/json
      description: FindById Size from the store
      parameters:
      - description: Size ID
        in: path
        name: sizeId
        required: true
        type: string
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SizeResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: FindById Size from the store
      tags:
      - Size
    put:
      consumes:
      - application/json
      description: Update Size from the store
      parameters:
      - description: Size Update
        in: body
        name: Size
        required: true
        schema:
          $ref: '#/definitions/models.SizeUpdate'
      - description: Size ID
        in: path
        name: sizeId
        required: true
        type: string
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SizeResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Update Size from the store
      tags:
      - Size
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"